+++ HISTORY.md
@@ -5,7 +5,7 @@
 
 - Support for bitwise operations `bitAnd`, `bitNot`, `bitOr`, `bitXor`,
   `leftShift`, `rightArithShift`, and `rightLogShift`. Thanks @BigFav.
-- Support for boolean operations `and`, `not`, `or`. Thanks @BigFav.
+- Support for boolean operations `and`, `not`, `or`, `xor`. Thanks @BigFav.
 - Support for `gamma` function. Thanks @BigFav.
 - Converting a unit without value will now result in a unit *with* value,
   i.e. `inch in cm` will return `2.54 cm` instead of `cm`.